requirements.txt
transpose-data
plotly
pandas
streamlit